thread fix i think

+++ vm/vm.h
@@ -2,8 +2,8 @@
 
 #include "libc.h"
 
-typedef int16_t vm_reg_t;
-typedef int16_t vm_opcode_t;
+typedef int32_t vm_reg_t;
+typedef int32_t vm_opcode_t;
 
 #include "obj.h"
